😍 make builtin2.c
